Added temporary old version
This is an old/pre-release version of Katie.
This version definitely has bugs, but can technically be used -- the primary known issue in this version is in the self-framing routine which for some links doesn't behave as it should. However it does work for most input. 
This bug has been fixed for the proper release of Katie, so I recommend waiting for that.